Convert 104 to 256

+++ contracts/vaults/RibbonThetaVault.sol
@@ -180,11 +180,11 @@ contract RibbonThetaVault is RibbonVault, OptionsThetaVaultStorage {
         require(amount > 0, "!amount");
         require(depositReceipt.round == currentRound, "Invalid round");
 
-        uint104 receiptAmount = depositReceipt.amount;
+        uint256 receiptAmount = depositReceipt.amount;
         require(receiptAmount >= amount, "Exceed amount");
 
         // Subtraction underflow checks already ensure it is smaller than uint104
-        depositReceipt.amount = uint104(uint256(receiptAmount).sub(amount));
+        depositReceipt.amount = uint104(receiptAmount.sub(amount));
         vaultState.totalPending = uint128(
             uint256(vaultState.totalPending).sub(amount)
         );
@@ -246,9 +246,9 @@ contract RibbonThetaVault is RibbonVault, OptionsThetaVaultStorage {
     function _closeShort(address oldOption) private {
         optionState.currentOption = address(0);
 
-        uint104 lockedAmount = vaultState.lockedAmount;
+        uint256 lockedAmount = vaultState.lockedAmount;
         vaultState.lastLockedAmount = lockedAmount > 0
-            ? lockedAmount
+            ? uint104(lockedAmount)
             : vaultState.lastLockedAmount;
 
         vaultState.lockedAmount = 0;
